Add support for type 7 (sending input).
This commit adds support for Gopher type 7 (full-text search).  In
this update, an -InputObject parameter has been added, so that data
can be easily appended to the URL and sent to the server.

Pipeline input is also supported.

In previous versions of this module, the user had the option to make
a URL-encoded query string on their own.  That accident will still
work.  However, it is recommended to use -InputObject, as that will
handle encoding and other sanity checks.