Update content types to characters, not numbers
Due to an oversight, this app had been using numeric content types
instead of ASCII content types, in violation of the Gopher specs.
For example, instead of using '1' (ASCII 49) as the content type
identifier for Gophermaps, this app may have been using 1 (0x01).
This has been corrected, and in case anything was relying on this,
we're bumping the version number up.

I also fixed a bad regex that would have caused .ps1xml files not
to be recognized as text.